requirements.txt
numpy
streamlit
opencv-python-headless
pillow
streamlit-drawable-canvas
